+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for efficient self-managed encoding of multiple meshes. Users may use the advanced Mesh API to acquire readable mesh data for multiple meshes at once and pass the data on to new `DracoEncoder.EncodeMesh` overloads that accept said `MeshData`.
-- Vertex attribute to draco ID map was added to `EncodeResult`
+- Vertex attributes information (draco identifier and dimensions) was added to `EncodeResult`
 ### Changed
 - Updated Burst dependency to version 1.8.4
 ### Fixed